+++ lib/state_manager/widgets/widgets.dart
@@ -1,3 +1,4 @@
 export 'deex_store.dart';
 export 'list_notifier.dart';
 export 'simple_builder.dart';
+export 'deex_consumer.dart';