+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- rename simple folder to widgets
+- add DeexConsumer (new reactive widget)
+- documentation improvements
+
 ## 0.0.1+2
 
 - documentation improvements